* Made ThrottleNode an object held by PCB, rather than ThrottleList which now holds ThrottleNode*. Moved Rate from PCB to ThrottleNode.
git-svn-id: svn://rakshasa.no/libtorrent/trunk/libtorrent@596 e378c898-3ddf-0310-93e7-cc216c733640

+++ src/torrent/rate.cc
@@ -44,7 +44,7 @@ namespace torrent {
 inline void
 Rate::discard_old() const {
   while (!m_container.empty() &&
-	 m_container.back().first < Timer::cache().seconds() - m_span) {
+	 m_container.back().first < Timer::cache().seconds() - (int32_t)m_span) {
     m_current -= m_container.back().second;
     m_container.pop_back();
   }
